SPI Driver transmit working, now need to work on cleaning it up and getting recieve working

+++ pico/src/ws2812b.rs
@@ -2,8 +2,8 @@ use embassy_rp::spi;
 
 const DRIVER_FREQUENCY: u32 = 8_500_000;
 pub const INFO_SIZE: usize = 24;
-const LOGIC_0: u8 = 0xE0;
-const LOGIC_1: u8 = 0xFC;
+const LOGIC_0: u16 = 0xE0;
+const LOGIC_1: u16 = 0xFC;
 pub const NUM_LEDS: usize = 12;
 pub const LED_INFO_SIZE: usize = 2 + (INFO_SIZE * NUM_LEDS);
 
@@ -14,10 +14,10 @@ pub fn get_addressable_led_config() -> spi::Config {
     config
 }
 
-fn rgb_to_logic_buffer(color_value: u32) -> [u8; INFO_SIZE] {
+fn rgb_to_logic_buffer(color_value: u32) -> [u16; INFO_SIZE] {
     // Get the red logic
     let mut mask: u32 = 0x00800000;
-    let mut logic_buffer_bits: [u8; INFO_SIZE] = [0; INFO_SIZE];
+    let mut logic_buffer_bits: [u16; INFO_SIZE] = [0; INFO_SIZE];
 
     for i in 0..8 {
         if color_value & mask > 0 {
@@ -51,9 +51,9 @@ fn rgb_to_logic_buffer(color_value: u32) -> [u8; INFO_SIZE] {
     logic_buffer_bits
 }
 
-// Count must me the number of leds * INFO_SIZE
-pub fn generate_addressable_led_buffer<const COUNT: usize>(color_values: &[u32]) -> [u8; COUNT] {
-    let mut addressable_led_buffer: [u8; COUNT] = [0; COUNT];
+// Count must be the number of leds * INFO_SIZE
+pub fn generate_addressable_led_buffer<const COUNT: usize>(color_values: &[u32]) -> [u16; COUNT] {
+    let mut addressable_led_buffer: [u16; COUNT] = [0; COUNT];
 
     let mut index: u32 = 2;
     for color in color_values.iter() {
